% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{CERMBlidar-deprecated}
\alias{CERMBlidar-deprecated}
\alias{get_stratum_cover}
\title{Deprecated functions in package \pkg{CERMBlidar}.}
\usage{
get_stratum_cover(rcounts, background = NA)
}
\description{
Deprecated functions in package \pkg{CERMBlidar}.
}
\section{\code{get_stratum_cover}}{

Please update your code to replace calls to \code{get_stratum_cover} with \code{\link{get_cover_quantiles}}.
}

