% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightline_functions.R
\name{check_flightline_orientation}
\alias{check_flightline_orientation}
\title{Check that the orientations of all flight lines match}
\usage{
check_flightline_orientation(las, ...)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{prepare_tile}.}

\item{...}{Additional arguments to pass to \code{get_flightline_info}
including \code{angular.tol}.}
}
\value{
\code{TRUE} if flight lines are consistent; \code{FALSE} otherwise.
}
\description{
This is a convenience function that calls \code{get_flightline_info}
with default arguments and checks that all flight lines have the same
orientation, either 'EW' or 'NS'.
}
\seealso{
\code{\link{get_flightline_info}}
}
