% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{check_strata}
\alias{check_strata}
\title{Check strata lookup table and put it into a standard form}
\usage{
check_strata(strata)
}
\arguments{
\item{strata}{A data frame of strata definitions.}
}
\value{
A data frame in standard form with columns name, lower, upper,
  and rows ordered by ascending stratum heights.
}
\description{
This function is used by \code{\link{get_stratum_counts}}. It checks that:
\itemize{
  \item The object is a data frame with (at least) three columns labelled:
    name, lower, upper. Column order and case of names does not matter.
  \item All stratum upper heights are greater than lower heights.
  \item No two strata overlap (gaps between strata are permitted).
}
An error results if any of these conditions are not met.
}
