% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_functions.R
\name{compile_metadata}
\alias{compile_metadata}
\title{Create a catalog of LAS tiles from HTML metadata}
\usage{
compile_metadata(path, dirs = "LAS", exts = "html",
  extent.crs = 4326)
}
\arguments{
\item{path}{Top of the directory tree containing tiles to catalog.}

\item{dirs}{One or more path elements to subset the directory tree. The
default is "LAS" to ignore, for example, "DEM" directories present in data
provided by Spatial Services, New South Wales. Set to NULL or an empty
string if not required.}

\item{exts}{One or more file extensions to identify metadata files.}

\item{extent.crs}{A coordinate reference system specifier (e.g. integer EPSG
code) to use for LAS tile extent polygons. The default is EPSG:4326
(WGS84). Set to \code{NULL} or \code{sf::NA_crs_} if you do not want to
apply a single CRS to all extent polygons.}
}
\value{
A spatial data frame containing summary metadata and tile extent
  polygons.
}
\description{
This function recurses through a directory tree containing (usually) LAS
tiles and associated HTML metadata documents, and compiles a summary in the
form of a spatial data frame (class \code{sf} object). The geometry column of
the spatial data frame consists of rectangular polygons for tile extents.
Presently, the function assumes that HTML metadata is in the format used by
Spatial Services, New South Wales.
}
\examples{
\dontrun{
# Compile metadata and save as an ESRI shapefile
meta <- compile_metadata("D:/mydata")
sf::st_write(meta, "metadata.shp", delete_layer = TRUE)
}

}
\seealso{
\code{\link{read_html_metadata}}
}
