% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{filter_flightlines}
\alias{filter_flightlines}
\title{Filter flight lines based on the number of points in each}
\usage{
filter_flightlines(las, min.points = 1000)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{prepare_tile}.}

\item{min.points}{The minimum number of points in a flight line for it to be
retained. The default value (1e3) is intended to exclude flight lines that
only appear at the margins of the tile.}
}
\value{
A (possibly) modified copy of the input tile with any sparse flight
  lines removed.
}
\description{
This function takes an input LAS tile and returns a copy from which any
flight lines with less than a specified minimum number of points have been
removed. It is used by \code{prepare_tile} but can also be called
directly. If points are dropped, the header of the returned LAS object is
updated.
}
