% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_bounding_rectangle}
\alias{get_bounding_rectangle}
\title{Find the minimum bounding rectangle that encloses specified points.}
\usage{
get_bounding_rectangle(las, classes = "all", flightlines = "all")
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{\link{prepare_tile}}.}

\item{classes}{Point classes to include: either a vector of integer class
numbers or the string \code{'all'} (default).}

\item{flightlines}{Flight lines to include: either a vector of integer
class numbers or the string \code{'all'} (default).}
}
\value{
A data frame of corner vertices for the minimum bounding rectangle.

Adapted from code by 'whuber' at: https://gis.stackexchange.com/a/22934/59514
}
\description{
When called with a single argument for the LAS tile, this function returns
the minimum bounding rectangle for the point cloud. The \code{classes} and
\code{flightlineIDs} arguments can be used to define a subset of points
to consider. If another kind of subset is required, the helper function
\code{link{.min_rectangle}} can be used directly.
}
