% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{get_class_frequencies}
\alias{get_class_frequencies}
\title{Get point class frequencies.}
\usage{
get_class_frequencies(las)
}
\arguments{
\item{las}{A LAS object.}
}
\value{
A named list of the count of points in each of the following
classes: ground, veg, building, water, other.
}
\description{
Gets the number of points in classes: ground, veg, building, water, and
other.
}
