% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_flightline_dir}
\alias{get_flightline_dir}
\title{Get the dimensions and orientations of flight lines}
\usage{
get_flightline_dir(las, ratio = 1.5)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{\link{prepare_tile}}.}

\item{ratio}{Minimum ratio of X and Y dimensions to be classified as either
E-W or N-S. Flight lines that do not satisfy this ratio are classified
as 'XX'.}
}
\value{
A data frame with columns: flightlineID, xlen, ylen and orientation.
}
\description{
Get the dimensions and orientations of flight lines
}
