% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{get_las_bounds}
\alias{get_las_bounds}
\title{Get the bounding rectangle of the point cloud}
\usage{
get_las_bounds(las, type = c("vec", "wkt", "sf"))
}
\arguments{
\item{las}{A LAS object.}

\item{type}{Either \code{'vec'} (default) to return a named vector of min
and max coordinates; \code{'wkt'} to return a WKT text string, or
\code{'sf'} to return a polygon object.}
}
\value{
The bounding polygon in the format specified by the \code{type}
  argument.
}
\description{
This function constructs a polygon based on the minimum and maximum X and Y
ordinates in the LAS data table, and returns it as either a named vector, a
WKT (Well Known Text) string specifier, or an \code{sf} polygon object.
}
