% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{get_point_density}
\alias{get_point_density}
\title{Calculate average point density for a tile}
\usage{
get_point_density(las)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{\link{prepare_tile}}.}
}
\value{
The average point density as a single value.
}
\description{
This function counts the points in the tile and divides by the total area
(based on the applicable distance units, e.g. square metres). For a tile with
average point density \code{p}, an estimate of the average distance between
neighbouring points, disregarding edge effects, is \code{0.5 / sqrt(p)}.
}
