% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{get_stratum_counts}
\alias{get_stratum_counts}
\title{Get LiDAR point counts for defined strata.}
\usage{
get_stratum_counts(las, strata, res = 10, classes = c(2, 3, 4, 5, 9))
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{prepare_tile}.}

\item{strata}{A data frame of strata definitions with (at least) three
columns: label, lowerht, upperht. Strata must cover all heights and be
non-overlapping. This will be checked by the function \code{check_strata}.}

\item{res}{Raster cell size.}

\item{classes}{Integer classification codes for points to include.
Default is 2 (ground) and 3:5 (vegetation) and 9 (water).}
}
\value{
A \code{terra::SpatRast} object where each band corresponds to a
  stratum (including the ground layer) and cell values are voxel point
  counts.
}
\description{
Give a LAS tile and a set of height ranges, this function creates a
multi-band raster where the cells in each band represent voxels for a given
stratum, and each cell value is the number of points within the voxel. This
can be used as the first step in the calculation of vegetation cover
estimates.
}
\seealso{
\code{\link{check_strata}}
}
