% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{get_stratum_cover-deprecated}
\alias{get_stratum_cover-deprecated}
\title{Calculate vegetation cover estimates for strata}
\usage{
get_stratum_cover(rcounts, background = NA)
}
\arguments{
\item{rcounts}{A raster object with a band of point count data for each
vegetation stratum plus a band for the ground layer. Normally, this will be
a \code{terra} package \code{SpatRaster} object generated by function
\code{get_stratum_counts()}, but it may also be a \code{RasterStack} or
\code{RasterBrick} (\code{raster} package) object.}

\item{background}{Value (default \code{NA}) to assign as the cover value for
a stratum when there are no LiDAR points for the stratum, any lower
strata or ground.}
}
\value{
A \code{SpatRast} object with a band for each vegetation stratum.
}
\description{
This function was formerly used to calculate maximum likelihood
  mean cover estimates for strata. Please now use function
  \code{\link{get_cover_quantiles}} instead with argument \code{mle=TRUE}.
}
\seealso{
\code{\link{CERMBlidar-deprecated}}
}
\keyword{internal}
