% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_stratum_cover}
\alias{get_stratum_cover}
\title{Calculate strata cover estimates}
\usage{
get_stratum_cover(rcounts)
}
\arguments{
\item{rcounts}{A \code{RasterStack} or \code{RasterBrick} with a layer for
each vegetation stratum plus a layer for ground. Cell values are point
counts.}
}
\value{
A \code{RasterStack}
}
\description{
This function takes a \code{RasterStack} of point count data for the ground
layer and vegetation strata and calculates corresponding layer cover
estimates. It assumes that the first layer is for ground points and
subsequent layers are for vegetation strata in increasing height order.
}
