% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{highest_density_interval}
\alias{highest_density_interval}
\title{Calculate highest density interval for a vector of values}
\usage{
highest_density_interval(x, prob = 0.95)
}
\description{
Calculate highest density interval for a vector of values
}
