% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{is_empty_tile}
\alias{is_empty_tile}
\title{Check whether a LAS tile has an empty point cloud}
\usage{
is_empty_tile(las)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{\link{prepare_tile}}.}
}
\value{
TRUE if the point cloud is empty.
}
\description{
Checks whether a point cloud is empty, i.e. there are zero rows in the LAS
object's data table.
}
