% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{is_prepared_tile}
\alias{is_prepared_tile}
\title{Check whether a LAS tile object has been prepared}
\usage{
is_prepared_tile(las)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{\link{prepare_tile}}.}
}
\value{
TRUE if the tile has been prepared or FALSE otherwise.
}
\description{
When a LAS tile is imported using the \code{\link{prepare_tile}} function,
point heights are normalized relative to ground elevation. This adds an
extra column (Zref) to the LAS data table.
}
