% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{layer_summary_metrics}
\alias{layer_summary_metrics}
\title{Metrics function for summary statistics}
\usage{
layer_summary_metrics(z, prob, breaks)
}
\description{
Metrics function for summary statistics
}
