% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{mandatory_input_fields}
\alias{mandatory_input_fields}
\title{Get names of mandatory input fields for LAS tiles}
\usage{
mandatory_input_fields()
}
\value{
A named vector of lower case field names where names are standard,
  single character field abbreviations.
}
\description{
Returns a named vector of character strings, where names are standard, single
character field abbreviations (as understood by \code{prepare_tile} and
\code{lidR::readLAS}) and values are field names. These are the fields that
must be included when importing a tile with \code{prepare_tile}.
}
\seealso{
\code{prepare_tile}
}
