% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{mask_tile}
\alias{mask_tile}
\title{Apply a raster mask to a LAS tile}
\usage{
mask_tile(las, r)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{\link{prepare_tile}}.}

\item{r}{A \code{RasterLayer} object to use as the mask.}
}
\value{
A new \code{LAS} object containing the masked points and updated
  header information.
}
\description{
This function uses a raster layer as a mask to subset a LAS point cloud.
Those points that fall within raster cells with non-missing values are
retained. Any points outside the bounds of the raster, or falling in cells
with \code{NA} values are dropped. If no points fall within data cells of
the raster, a warning message is issued and \code{NULL} is returned.
}
\note{
This function \strong{does not check} that the coordinate reference
  systems of the LAS and raster objects are the same.
}
