% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flightline_functions.R
\name{plot_flightline_points}
\alias{plot_flightline_points}
\title{Plot a sample of points using colours to indicate flight lines}
\usage{
plot_flightline_points(
  las,
  classes = c(2, 35),
  npts = 5000,
  shape = 16,
  size = 1
)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{prepare_tile}.}

\item{classes}{An integer vector specifying the point classes to display. May
include 2 (ground), 3 (low veg), 4 (mid veg), 5 (high veg), 35 (all veg -
the three vegetation classes combined), 6 (buildings) and 9 (water). The
default is to display ground (2) and all vegetation combined (35).}

\item{npts}{Number of points to sample from the LAS data for drawing.
Sampling is done proportionally across combinations of point classes and
flight lines, and the resulting number of rows in the \code{data} element
of the returned ggplot object can vary slightly. If displaying many point
classes, increasing this value from the default (5000) and decreasing the
point size will produce nicer plots.}

\item{shape}{Point shape, specified as an integer code as for gpplot and
R base plot.}

\item{size}{Point size.}
}
\value{
A ggplot object.
}
\description{
Displays a random sample of points from a LAS object, coloured by flight line
ID and facetted by point class. This is useful, for example, when diagnosing
problems with overlapping flightlines or uneven point density.
}
\examples{
\dontrun{
# Default plot of points for ground (class 2) and combined
# vegetation (classes 3-5).
plot_flightline_points(las)

# Display the vegetation classes both separately and combined
plot_flightline_points(las, classes = c(3:5, 35))
}

}
\seealso{
\code{\link{get_flightline_polygons}}
}
