% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog_functions.R
\name{read_html_metadata}
\alias{read_html_metadata}
\title{Read metadata for a LAS tile}
\usage{
read_html_metadata(path.html)
}
\arguments{
\item{path.html}{The full path to an HTML document containing tile metadata.}
}
\value{
A named list with the following elements:
  \describe{
    \item{filename}{Metadata file name.}
    \item{locality}{Place name (from file name).}
    \item{lidnum}{LID or PHO number, e.g. 'PHO3' (from file name).}
    \item{classnum}{Class number, e.g. 'C0' (from filename).}
    \item{ahdnum}{AHD number, e.g. 6386236 (from filename).}
    \item{startdate}{Start date of data capture (from doc).}
    \item{enddate}{End date of data capture (from doc).}
    \item{bounds.lon}{Vector of min and max longitude (from doc).}
    \item{bound.lat}{Vector of min and max latitude (from doc).}
    \item{epsg}{EPSG code for tile (from doc). To be used if converting to UTM coordinates.}
    \item{datum}{Horizontal datum, e.g. GDA94 (from doc).}
  }
}
\description{
This function reads metadata from an HTML document in the format used by
Spatial Services, New South Wales (Australia). It also extracts attributes
from the document file name. It is used by function \code{compile_metadata}
but can also be called directly.
}
\seealso{
\code{\link{compile_metadata}}
}
