% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{relate_strata}
\alias{relate_strata}
\title{Relate two definitions of strata}
\usage{
relate_strata(strata1, strata2)
}
\arguments{
\item{strata1}{A data frame of strata definitions. Usually this should be
the set with the smaller number of strata.}

\item{strata2}{A data frame of strata definitions. Usually this should be
the set with the larger number of strata.}
}
\value{
A named list, where element names are taken from \code{strata1}
  and each element is a vector of the indices of the related strata in
  \code{strata2}.
}
\description{
Takes two strata definitions (e.g. StrataSpecht and StrataCERMB) and returns
a list where the number of elements is equal to the number of strata in the
first set, and each element is a vector with indices for strata in the
second set.
}
