% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_functions.R
\name{remove_flightline_overlap2}
\alias{remove_flightline_overlap2}
\title{Remove overlap between flight lines using a raster approach}
\usage{
remove_flightline_overlap2(las, res = 10, classes = 5)
}
\description{
Remove overlap between flight lines using a raster approach
}
