% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{remove_zero_gpstimes}
\alias{remove_zero_gpstimes}
\title{Check for and drop any points with zero GPS times}
\usage{
remove_zero_gpstimes(las, max_prop = 0.01, quiet = FALSE)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{prepare_tile}.}

\item{max_prop}{The maximum proportion of points with zero GPS times. This
should be set to a small value (default is 0.01). If the proportion of zero
GPS times exceeds this value, the function will throw an error.}

\item{quiet}{Set to \code{TRUE} to suppress messages. Default is
\code{FALSE}.}
}
\value{
The (possibly modified) LAS object.
}
\description{
Sometimes the GPS time for a point is zero rather than the expected large
integer value (seconds since epoch), which will lead to incorrect capture
times being reported for the LiDAR tile. This function checks for the
presence of such invalid GPS times and drops those points from the data if
they constitute no more than a specified proportion of the total points. The
function returns the (possibly modified) LAS object.
}
\details{
Note: Any points with a missing value for GPS time will be retained.
}
\examples{
\dontrun{
las <- remove_zero_gpstimes(las)
}

}
\seealso{
\code{\link{get_scan_times}}
}
