% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{reproject_tile}
\alias{reproject_tile}
\title{Transform point cloud coordinates to a new projection}
\usage{
reproject_tile(las, epsg.new, epsg.old = lidR::epsg(las))
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{\link{prepare_tile}}.}

\item{epsg.new}{Integer EPSG code for the projection into which the point
cloud will be transformed.}

\item{epsg.old}{This argument allows you to optionally supply or override the
EPSG code for the existing projection. For a LAS tile with a properly
assigned projection this defaults to the existing EPSG code.}
}
\value{
A new \code{LAS} object with reprojected point coordinates and
  an updated header.
}
\description{
This function takes the X,Y coordinates of all points and transforms them
from their existing projection to a user-specified one. The transformation is
done using \code{sf} package methods. If the point cloud is very large this
may take a while.
}
