% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StrataCERMB}
\alias{StrataCERMB}
\title{Strata definitions for CERMB use.}
\format{
A data frame (tbl_df) with 61 rows and 3 columns:
\describe{
\item{name}{Stratum label ('to0.3', 'to0.5', 'to1.0' ... 'over30.0')}
\item{lower}{Lower height of stratum (metres)}
\item{upper}{Upper height of stratum (metres)}
}
}
\usage{
StrataCERMB
}
\description{
The lowest two strata are 0-0.3m and 0.3-0.5m. Subsequent strata are
0.5m intervals up to 10m; then 1m intervals up to 30m; then 5m
intervals up to 80m; then a catch-all layer for >80m.
}
\keyword{datasets}
