% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StrataCERMB_OLD}
\alias{StrataCERMB_OLD}
\title{Former strata definitions used by CERMB.}
\format{
A data frame with 6 rows and 5 columns:
\describe{
\item{strata_def}{Name for this strata scheme ('cermb_old')}
\item{index}{Integer index, ordered by height level}
\item{label}{Stratum label (Ground, 'upto0.5m', 'upto1m' ... 'over30m')}
\item{lowerht}{Lower height of stratum (metres)}
\item{upperht}{Upper height of stratum (metres)}
}
}
\usage{
StrataCERMB_OLD
}
\description{
Strata formerly used within the Centre for Environmental Risk Management of
Bushfires, University of Wollongong, to rasterize LiDAR point clouds. These
have now been replaced by \code{\link{StrataCERMB}}. The lowest two strata
are 0-0.3m and 0.3-0.5m. Subsequent strata are 0.5m intervals up to 30m; then
a catch-all layer for >30m.
}
\keyword{datasets}
