% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StrataSpecht}
\alias{StrataSpecht}
\title{Strata definitions based on Specht}
\format{A data frame (tbl_df) with 5 rows and 3 columns:
\describe{
\item{name}{Stratum name (LowShrub, TallShrub, LowTree, MediumTree, TallTree)}
\item{lower}{Lower height of stratum (metres)}
\item{upper}{Upper height of stratum (metres)}
}}
\usage{
StrataSpecht
}
\description{
Strata definitions based on Specht
}
\keyword{datasets}
