% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{check_flightlines}
\alias{check_flightlines}
\title{Check that the orientations of all flightlines match}
\usage{
check_flightlines(las, ...)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{\link{prepare_tile}}.}

\item{...}{Additional arguments to pass to \code{\link{get_flightline_info}}
including \code{angular.tol}.}
}
\value{
\code{TRUE} if flight lines are consistent; \code{FALSE} otherwise.
}
\description{
This is a convenience function that calls \code{\link{get_flightline_dir}}
with default arguments and checks that all flight lines have the same
orientation, either 'EW' or 'NS'.
}
\seealso{
\code{\link{get_flightline_info}}
}
