% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_cover_functions.R
\name{get_cover_difference}
\alias{get_cover_difference}
\title{Calculate quantiles for change in vegetation cover between two times}
\usage{
get_cover_difference(
  rcounts0,
  rcounts1,
  probs = c(0.05, 0.5, 0.95),
  backgroundNA = FALSE,
  beta_prior = c(1, 1)
)
}
\arguments{
\item{rcounts0}{Point counts raster for the first (earlier) time. Normally,
this will be a \code{terra} package \code{SpatRaster} object generated by
function \code{get_stratum_counts()}, but it may also be a
\code{RasterStack} or \code{RasterBrick} (\code{raster} package) object.
\strong{Important:} the bands must be arranged in increasing height order,
i.e. ground layer, then understorey layer(s), then overstorey layer(s).}

\item{rcounts1}{Point counts raster for the second (later) time. This must
have the same dimensions (rows, columns and number of bands) as
\code{rcounts0}. Generally it will also have the same spatial bounds, but
this is not checked to allow the function to be used for other types of
comparisons (e.g. space for time substitution).}

\item{probs}{A numeric vector of one or more probability values specifying
the quantiles of vegetation cover change to calculate.  The default
\code{c(0.05, 0.5, 0.95)} returns quantiles for the median and the central
90\% interval of cover change.}

\item{backgroundNA}{Controls what pixel value should be returned for a
stratum when there are no LiDAR points for the stratum , any lower strata
or ground level in one or both point counts rasters. If set to \code{TRUE},
such cases will have missing (\code{NA}) pixel values for all quantiles for
the stratum. If \code{FALSE} (the default), quantiles for change values
will be based on the prior beta distribution.}
}
\value{
A \code{SpatRast} (\code{terra} package) raster object with a layer
  for each requested quantile of vegetation cover change within each stratum.
  For example, calculating the default median plus 90\% bounds (0.05 and 0.95
  quantiles) of cover change for 5 vegetation strata would return a raster
  object with 15 layers. Layers are arranged by stratum, then by quantile.
  Layer names have the form \code{stratum_dq_prob}, e.g.
  \code{'TallShrub_dq_0.5'} for the median (50\%) quantile of cover change.
}
\description{
This function uses the same approach for Bayesian estimation of vegetation
cover as described for function \code{\link{get_cover_quantiles}}, but
applied to two rasters representing LiDAR point counts for the same area at
two different times. It returns a raster where the values in each band are
selected quantiles of the change in vegetation cover between time 1 and time
2. The default is to calculate the median and the central 90\% bounds (i.e.
5\%, 50\% and 95\% quantiles) for cover change in each stratum. Each input
raster should have two or more bands, with the first band representing ground
point counts and subsequent bands representing point counts for vegetation
strata arranged in increasing height order (i.e. ground is band 1).
}
\details{
As described for function \code{\link{get_cover_quantiles}}, the estimation
of stratum vegetation cover is based on treating the discrete LiDAR returns
as a binomial sampling process, where the probability of a return from
vegetation in the stratum being considered corresponds to vegetation cover.
Under this model, the distribution of possible cover values that could have
generated the observed point counts for a pixel will follow a beta
distribution. Change in vegetation cover can therefore be estimated by taking
the distribution of differences between the beta distribution for time 2
minus that for time 1.
}
