% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get_flightline_info}
\alias{get_flightline_info}
\title{Get summary information for flight line extents}
\usage{
get_flightline_info(las, ratio = 1.5)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{\link{prepare_tile}}.}

\item{ratio}{Minimum ratio of X and Y dimensions to be classified as either
E-W or N-S. Flight lines that do not satisfy this ratio are classified
as 'XX'.}
}
\value{
A spatial data frame (class \code{sf}) with columns: flightlineID,
  xlen, ylen, orientation and geometry (minimum bounding rectangle of flight
  line).
}
\description{
This function determines the minimum bounding rectangle of points in each flight line
and, based on this, the orientation: one of NS (north - south), EW (east - west),
or XX (indeterminate).
}
\seealso{
\code{\link{check_flightlines}}
}
