% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{get_horizontal_crs}
\alias{get_horizontal_crs}
\title{Get the horizontal coordinate reference system}
\usage{
get_horizontal_crs(las)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{prepare_tile}.}
}
\value{
An \code{sf} package \code{crs} object.
}
\description{
Recent LiDAR data (e.g. from 2020 in Australia) can have a compound
coordinate reference system defined that consists of a horizontal and a
vertical component. Usually, we just want to use the horizontal component for
derived raster layers. This function extracts the horizontal component and
returns it as a \code{crs} object as used by the \code{sf} package. If the
input object has a simple (i.e. non-compound) coordinate reference syste,
that will be returned unchanged.
}
