% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{get_max_height}
\alias{get_max_height}
\title{Derive a raster of maximum point height in all or selected classes}
\usage{
get_max_height(las, res = 10, classes = c(2, 3:5, 9), nodata = NA)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{prepare_tile}.}

\item{res}{Raster cell size. The default value is 10 which assumes that
the LAS data is in a projected coordinate system with metres as units.}

\item{classes}{Point classes to include when calculating maximum height
for each cell. Default is c(2, 3:5, 9) for ground, vegetation and
water point classes.}

\item{background}{Integer value for cells with no points. Default is \code{NA}.}
}
\value{
A \code{RasterLayer} in which cell values are maximum point height.
}
\description{
This is useful for checking purposes. It can also be used to derive
an approximate vegetation height layer by setting the \code{classes}
argument to \code{c(3:5)}.
}
