% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{get_stratum_cover}
\alias{get_stratum_cover}
\title{Calculate vegetation cover estimates for strata}
\usage{
get_stratum_cover(rcounts)
}
\arguments{
\item{rcounts}{A raster object with a band of point count data for each
vegetation stratum plus a band for the ground layer. Normally, this will be
a \code{terra} package \code{SpatRast} object generated by function
\code{get_stratum_counts()}, but it may also be a \code{RasterStack} or
\code{RasterBrick} (\code{raster} package) object.}
}
\value{
A \code{SpatRast} object with a band for each vegetation stratum.
}
\description{
This function takes a raster of point count data for specified strata
(usually generated by function \code{get_stratum_counts}) and calculates
corresponding cover estimates. It assumes that the first layer is for ground
points and subsequent layers are for vegetation strata in increasing height
order.
}
