% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{plot_flightlines}
\alias{plot_flightlines}
\title{Quick plot of flight lines}
\usage{
plot_flightlines(las, npts = 5000, shape = 16, size = 1)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{\link{prepare_tile}} or
\code{\link[lidR]{readLAS}}.}

\item{npts}{Number of points to sample from the LAS object (default: 5000).}

\item{shape}{Point shape, specified as an integer code as for gpplot and
R base plot.}

\item{size}{Point size.}
}
\description{
Displays points locations for a random sample of points from a LAS object,
with points coloured by flight line ID. This is useful for detecting
overlapping flightlines.
}
