% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{plot_flightlines}
\alias{plot_flightlines}
\title{Plot point locations and flight lines}
\usage{
plot_flightlines(las, npts = 5000, shape = 16, size = 1)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{\link{prepare_tile}}.}

\item{npts}{Number of points to draw (default: 5000). The subset of
points is selected to cover all combinations of flight line and
point class.}

\item{shape}{Point shape, specified as an integer code as for gpplot and
R base plot.}

\item{size}{Point size.}
}
\value{
A ggplot object.
}
\description{
Displays points locations for a random sample of points from a LAS object,
with points coloured by flight line ID. This is useful for detecting
overlapping flightlines.
}
\examples{
\dontrun{
# Quick plot
plot_flightlines(las)

# Facet by point classes
plot_flightlines(las) + facet_wrap(~ Classification)
}

}
