% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{prepare_tile}
\alias{prepare_tile}
\title{Import and prepare a LAS tile for further processing}
\usage{
prepare_tile(path, drop.negative = TRUE, fields = "*", classes = c(2,
  3, 4, 5, 6), flight.gap = 60, remove.overlap.class = 5)
}
\arguments{
\item{path}{Path to the LAS tile to process.}

\item{drop.negative}{If TRUE, any points (other than ground points) whose
heights are below ground level (as estimated by Delaunay interpolation of
ground point heights) are discarded.}

\item{fields}{A string containing single-letter abbreviations for the data
fields to include. The default is to return all fields. See
\code{\link[lidR]{readLAS}} for details of the available abbreviations.}

\item{classes}{Point classifcations to include. Default is ground (2), vegetation
(3, 4, 5) and buildings (6).}

\item{flight.gap}{The minimum time gap (seconds) to use when assigning points
to flight lines.}

\item{remove.overlap.class}{A vector of one or more point classes (default: 5
for high vegetation) Any overlap between flight lines is removed for the
specified classes by calling \code{\link{remove_flightline_overlap}}.
To disable, set this argument to an empty vector (\code{integer(0)}), or
\code{NA} or \code{NULL}.}
}
\value{
A \code{LAS} object.
}
\description{
This function imports data from a LAS file and prepares it for further
processing. A surface model is first fitted to ground points (class 2) by
Delaunay triangulation and the elevation of all points is then adjusted to be
relative to ground level. Any areas of overlap between flight lines (e.g.
seen for vegetation class 5 points in NSW Government data) are removed by
default.
}
