% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{prepare_tile}
\alias{prepare_tile}
\title{Import and prepare a LAS tile for further processing}
\usage{
prepare_tile(
  path,
  normalize.heights = "tin",
  treat.as.ground = c(2, 9),
  drop.negative = TRUE,
  drop.negative.threshold = 0,
  fields = NULL,
  classes = NULL,
  min.points = 1000,
  flight.gap = 60,
  unzip.dir = NULL
)
}
\arguments{
\item{path}{Path to the LAS or LAZ format file to process. If the file
extension is '.zip' it is assumed to be a compressed LAS file that will be
unzipped before processing (see the \code{unzip.dir} parameter below). A
compressed file should contain only one LAS file (identified by having a
'las' or 'LAS' file extension) although it can also contain other files
(e.g. HTML or XML documents).}

\item{normalize.heights}{Whether, and how, to normalize point heights to
ground level. Can be one of the following:
\describe{
  \item{A logical value}{If \code{TRUE}, relative ground level is estimated from
    an elevation surface fitted by triangulation to ground points in the
    LAS file. This is equivalent to calling the
    \code{\link[lidR]{normalize_height}} function directly with the argument
    \code{algorithm = tin()}. If \code{FALSE}, point heights will not be
    normalized.}
  \item{An algorithm name as a character string}{Point heights will be
    normalized using the specified algorithm. Must be one of:
    \code{'tin', 'knnidw', 'kriging'} which correspond to the algorithm
    functions provided by the \code{lidR} package.}
  \item{A raster layer}{If a raster layer is provided, the cell values
    will be used as ground elevation to normalize point heights. The layer
    should have the same or greater extent as the LAS file.}
  \item{A raster filename}{Any character string that does not match one of
    the supported algorithm names will be treated as the path to a raster
    file. Supported formats are GeoTIFF ('.tif') and ESRI ASCII ('.asc').
    If the file extension is '.zip' it is assumed to be a compressed
    file that will be unzipped before processing (see the \code{unzip.dir}
    parameter below). A compressed file should contain only one raster file
    (identified by having a '.tif' or '.asc' file extension) although it
    can also contain other files (e.g. HTML or XML documents).}
  \item{NULL}{Same as \code{FALSE}, ie. point heights will not be
    normalized.}
}
\strong{The default value is \code{'tin'}.}
If point heights are normalized, the original values are copied to a new
data table column: 'Zref'.}

\item{treat.as.ground}{Point classes to treat as ground points when
normalizing point heights by interpolation using one of the lidR package
algorithms (\code{'tin', 'knnidw', 'kriging'}). If point heights are being
normalized from a raster DEM, any classes specified by this argument will
have their normalized heights set to zero. The default value \code{c(2,9)}
specifies ground and water classes.}

\item{drop.negative}{If \code{TRUE} AND point heights are being normalized,
any points whose heights are more than \code{drop.negative.threshold} below
ground level will be discarded.}

\item{drop.negative.threshold}{If \code{drop.negative} is \code{TRUE} AND
point heights are being normalized, any heights more than this threshold
value below ground level will be discarded. Default value is zero.}

\item{fields}{Either \code{NULL} (default) to include all data fields, or a
character string containing single-letter abbreviations for selected
fields. See \code{\link[lidR]{readLAS}} for details of the available,
single-letter field abbreviations.}

\item{classes}{Point classes to include or exclude. The default
(\code{NULL}) is to include classes 2 (ground), 3-5 (vegetation), 6
(buildings) and 9 (water). Specify a subset of classes as a vector of
integers, e.g. \code{classes = 2:6} would include ground (2), vegetation
(3, 4, 5) and building (6) points. Negative values can be used to exclude
selected classes, e.g. \code{classes = -6} would include all classes except
those classified as building points. Note that overlap points (class 12)
are always excluded unless an explicit integer \code{classes} vector is
provided that includes the value 12.}

\item{min.points}{The minimum number of points in a flight line for it to be
retained in the imported tile. The default value (1000) is intended to
exclude flight lines that only appear at the margins of the tile.}

\item{flight.gap}{The minimum time gap (seconds) to use when assigning points
to flight lines.}

\item{unzip.dir}{The directory in which to uncompress a compressed LAS file
(identified by a '.zip' extension). If \code{NULL} (default) a temporary
directory will be used. After processing, the uncompressed file will be
deleted.}
}
\value{
A \code{LAS} object.
}
\description{
This function imports data from a LAS file and prepares it for further
processing. Point heights are normalized to relative height above ground,
either by fitting a surface model to ground points (plus water points by
default) or by taking ground elevations from a provided raster DEM. Height
normalization can be disabled if desired. Flight lines are identified based
on GPS times for points and, optionally, any flight lines with less than a
threshold number of points are discarded.
}
