% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{remove_flightline_overlap}
\alias{remove_flightline_overlap}
\title{Remove overlap between flight lines.}
\usage{
remove_flightline_overlap(las, classes = 5, res = 10, buffer = 100)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{\link{prepare_tile}}.}

\item{classes}{Vector of one or more integer codes for point classes to consider.}

\item{res}{Raster cell size to use when delineating overlap areas.}

\item{buffer}{Width of a buffer placed around the tile to ensure that the
boundaries partitioning overlap areas extend beyond all points.}
}
\value{
A modified copy of the input LAS object.
}
\description{
This function identifies areas of overlap between pairs of flight lines,
places a boundary along (approximately) the middle of the overlap area, and
removes points from each side that belong to the flight line (mainly) on the
other side. Before searching for, and removing, overlaps the function checks
that all flight lines have consistent orientation, ie. either all are
north-south or all are east-west. If this is not the case a warning message
is issued and the original, unmodified tile is returned.
}
\details{
We added this function because LIDAR data provided for New South Wales in
2018 had overlap between flight lines removed for all point classes other
than class 5 (high vegetation).
}
\seealso{
\code{\link{check_flightlines}}, \code{\link{plot_flightlines}}
}
