% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StrataCERMB}
\alias{StrataCERMB}
\title{Mostly regular strata from 0 - 30m, plus a catch-all stratum for >30m.
The lowest two strata are 0-0.3m and 0.3-0.5m. Subsequent strata are
0.5m intervals up to 30m.}
\format{A data frame (tbl_df) with 61 rows and 3 columns:
\describe{
\item{name}{Stratum label ('to0.3', 'to0.5', 'to1.0' ... 'over30.0')}
\item{lower}{Lower height of stratum (metres)}
\item{upper}{Upper height of stratum (metres)}
}}
\usage{
StrataCERMB
}
\description{
Mostly regular strata from 0 - 30m, plus a catch-all stratum for >30m.
The lowest two strata are 0-0.3m and 0.3-0.5m. Subsequent strata are
0.5m intervals up to 30m.
}
\keyword{datasets}
