% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StrataCERMB}
\alias{StrataCERMB}
\title{Strata definitions used by CERMB.}
\format{
A data frame with 6 rows and 5 columns:
\describe{
\item{strata_def}{Name for this strata scheme ('cermb')}
\item{index}{Integer index, ordered by height level}
\item{label}{Stratum label (Ground, 'upto0.5m', 'upto1m' ... 'over80m')}
\item{lower}{Lower height of stratum (metres)}
\item{upper}{Upper height of stratum (metres)}
}
}
\usage{
StrataCERMB
}
\description{
Strata used within the Centre for Environmental Risk Management of Bushfires,
University of Wollongong, to rasterize LiDAR point clouds. The lowest two
strata are 0-0.3m and 0.3-0.5m. Subsequent strata are 0.5m intervals up to
10m; then 1m intervals up to 30m; then 5m intervals up to 80m; then a
catch-all layer for >80m.
}
\keyword{datasets}
