% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StrataOFH}
\alias{StrataOFH}
\title{Strata definitions used for Overall Fuel Hazard Assessment in New South Wales.}
\format{
A data frame with 5 rows and 5 columns:
\describe{
\item{strata_def}{Name for this strata scheme ('ofh')}
\item{index}{Integer index, ordered by height level}
\item{label}{Stratum name (Ground, NearSurface, Elevated, LowCanopy, UpperCanopy)}
\item{lowerht}{Lower height of stratum (metres; -999.0 for ground layer)}
\item{upperht}{Upper height of stratum (metres; 999.0 for top layer)}
}
}
\usage{
StrataOFH
}
\description{
Strata definitions used for Overall Fuel Hazard Assessment in New South Wales.
}
\keyword{datasets}
