% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{StrataSpecht}
\alias{StrataSpecht}
\title{Strata definitions adapted from Specht 1970, 1974.}
\format{
A data frame with 6 rows and 5 columns:
\describe{
\item{strata_def}{Name for this strata scheme ('specht')}
\item{index}{Integer index, ordered by height level}
\item{label}{Stratum name (Ground, LowShrub, TallShrub, LowTree, MediumTree, TallTree)}
\item{lowerht}{Lower height of stratum (metres; -999.0 for ground layer)}
\item{upperht}{Upper height of stratum (metres; 999.0 for top layer)}
}
}
\usage{
StrataSpecht
}
\description{
Strata definitions adapted from Specht 1970, 1974.
}
\keyword{datasets}
