% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_functions.R
\name{update_tile_header}
\alias{update_tile_header}
\title{Update a LAS header based on current point data}
\usage{
update_tile_header(las)
}
\arguments{
\item{las}{A LAS object, e.g. imported using \code{prepare_tile}.}
}
\value{
A LAS object with header information updated, if necessary, to
  accord with its current data.
}
\description{
This function can be called to update the header information of a LAS object
after making direct changes to its points data (e.g. removing selected
points). It does a little more than the similar function \code{header_update}
in the \code{rlas} package which only updates the header elements for min and
max X, Y and Z, and the number of points. This function additionally updates
the X and Y offsets.
}
